Seis 52

+++ app/models/crew/crew_model.py
@@ -6,8 +6,10 @@ class Crew(TimestampMixin, ModelMixin, db.Model):
 
     __tablename__ = "crew"
 
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     name = db.Column(db.Text, nullable=False)
     shift = db.Column(db.Text, nullable=False)
-    role = db.Column(db.Enum('admin', 'manager', 'engineer'), nullable=False, default="N")
+    role = db.Column(
+        db.Enum("admin", "manager", "engineer"), nullable=False, default="N"
+    )
     manager_id = db.Column(db.Integer)
