+++ app/models/location.py
@@ -6,7 +6,7 @@ class LocationInfo(TimestampMixin, ModelMixin, db.Model):
 
     __tablename__ = "location_info"
 
-    id = db.Column(db.Integer, primary_key=True, nullable=False)
+    id = db.Column(db.BigInteger, primary_key=True, nullable=False)
     county_name_id = db.Column(db.Integer, nullable=False)
     basin_name_id = db.Column(db.Integer, nullable=False)
     state_id = db.Column(db.Integer, nullable=False)
@@ -38,7 +38,7 @@ class CountyName(TimestampMixin, ModelMixin, db.Model):
 
     __tablename__ = "county_name"
 
-    id = db.Column(db.Integer, primary_key=True, nullable=False)
+    id = db.Column(db.BigInteger, primary_key=True, nullable=False)
     county_name = db.Column(db.Text)
 
 
@@ -46,7 +46,7 @@ class BasinName(TimestampMixin, ModelMixin, db.Model):
 
     __tablename__ = "basin_name"
 
-    id = db.Column(db.Integer, primary_key=True, nullable=False)
+    id = db.Column(db.BigInteger, primary_key=True, nullable=False)
     basin_name = db.Column(db.Text, nullable=False)
 
 
@@ -54,5 +54,5 @@ class State(TimestampMixin, ModelMixin, db.Model):
 
     __tablename__ = "state"
 
-    id = db.Column(db.Integer, primary_key=True, nullable=False)
+    id = db.Column(db.BigInteger, primary_key=True, nullable=False)
     value = db.Column(db.Text, nullable=False)
